Utility/requirements.txt
streamlit
firebase-admin
pandas
numpy
pillow
requests
altair
pytz
plotly
